implemented DB update via GUI
