+++ sbMACROv2.0/app/main/templates/index.html
@@ -10,4 +10,6 @@
     // Set navigation bar to highlight current page
     let currentPage = document.getElementById('Nav-Home');
     currentPage.classList.add("currentPage");
-</script> {% endblock %}
+</script> 
+
+{% endblock %}
